refactor: move the scheduled refresh pass onto the provider interface
scheduled-token-refresh.ts hardcoded an OSF pass followed by a gdrive
pass. Each provider's pass now lives on its own adapter as an optional
refreshExpiringTokens(), and the scheduled function is a loop over the
registry that never names a provider. A provider that omits the method is
simply skipped.

The window stays provider-owned and the two are NOT interchangeable:
OSF's default is 2 weeks against refresh-token expiry
(refreshTokenExpires, `<=`), gdrive's is 10 minutes against access-token
expiry (tokenExpiresAt, `<`). Both queries, all log strings, the
no-refreshToken skip branch and the success/fail tallies move verbatim.
Registration order (osf, then gdrive) preserves the historical run order.

One deliberate behavior change, commented at the call site: the OSF pass
used to be unwrapped, so a hard failure threw out of the scheduled
function and the run was marked failed, while the gdrive pass was wrapped
and swallowed. A uniform loop has to pick one, so every provider is now
wrapped and logged, with no aggregate rethrow. That buys isolation --
one provider's failure can't stop another's pass -- at the cost of a
failing OSF pass no longer failing the scheduled run. If we'd rather have
the alert, collecting errors and throwing after the loop is a small
follow-up.

oauth-connect-refresh-emulator.test.js is re-pointed from the removed
refreshExpiringGdriveTokens export to gdriveProvider.refreshExpiringTokens
with no assertion changes, plus the node-fetch stub the other adapter
suites already carry. oauth-connect-scheduled-regression.test.js, which
POSTs the scheduled function in the emulator, is untouched and still
passes -- it is the regression guard for this change.

With this, adding a provider touches only its adapter, one
registerProvider line, and the StorageProviderId union: resolve-token,
oauth config and scheduled refresh are all closed to modification.
